Refactor `wasmtime-component-macro` a bit
Make room for a `wit_bindgen` macro to slot in.
